agent/requirements.txt
livekit
livekit-api
livekit-agents
livekit-plugins-openai
livekit-plugins-groq
livekit-plugins-silero
python-dotenv>=0.19.0
requests>=2.26.0
sentence-transformers
faiss-cpu
opencv-python-headless
Pillow